PortScannerDetector
* Modularized functions into their own files
* Using default python dictionary instead of computing index manually
* Using channels among threads to pass data in a threadsafe way
* Removed timer thread (put into detector thread function)
* Added dissector thread so that the sniffer won't drop packets while dissecting
* Tested running of program across a span of 5 minutes for a -T4 speed on nmap
* Made [Ctrl+C] exit the program smoothly (as opposed to needing to Ctrl+C for each thread)

At this point, the dst_ip is not used. It is not clear if it could be used.
This is something to test across at least 3 machines on a network.

I think different speeds should be tested, and it would be interesting to
even test multiple port scanners at once. At the moment, the port scanner
still needs to be updated to scan at different rates, since in the end it
seems we should be using that over nmap.

Note: I also removed comments temporarily. This is why I haven't deleted
the original file yet. They can be added back, but I didn't want to since
I would have felt compelled to make similar comments for new functions.
TL;DR The result of this is a combination of OCD and Laziness