Added promiscuous scan
* Added new file that can make a promiscuous socket
* Update TCP dissect packet to extract flag information
* Filtered out non-syn packets

Note: At this moment, the detector cannot a scanner on the
host. At one point I thought it could. Some modifications
to the flags may be needed. When running nmap it seems we
use a network hop and nmap says the packets are filtered.
I did not notice this for other nmap runs.